Silence cut performance fix
replace communication from cep to jsx, now is event driven for user updates.
thanx for that the loop is direct and varaible are cached, reduceing the inforamtion needed from premiere it self(slows down) AND fixed a bug when deleting audio track it rippled delete (there is no need to) which is slower then normal delete (more noticeable with multiple audis).

also minors fixes and attempt to remove any code that is not a must.

dwbugging side - added a timeout for ffmpeg io in case of freeze from output so i could debug it from users.