SilenceCut timeout handling
sometimes a freeze in std will trigger when the std is in "running" state but no output is provided making it wait for output even when std is closed.

this will better handle timeout and try to keep it running after that.
Also a minor bug fix that case a wrong timing error when a 1 frame cut is needed and it gets rounded wrongly